requirements.txt
fastapi~=0.63.0
pillow~=8.1.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
python-multipart==0.0.5
aiofiles==0.4.0
pycocotools>=2.0.1
opencv-python==4.3.0.36
detectron2~=0.3+cpu
numpy~=1.20.1
starlette~=0.13.6